+++ fastcookapp/static/fastcookapp/newGraph.js
@@ -17,24 +17,35 @@ function getCookie(name) {
 
 
 function doFunction(id,tr){
-    csrftoken = getCookie('csrftoken');
-    event.preventDefault();
-    /*var id = $(this).attr('id');
-    var tr = $(this).closest('tr');*/
-    console.log(id)
-    $.ajax({
-        type: "DELETE",
-        url: id, 
-        headers:{
-                "X-CSRFToken": csrftoken
-                },
-        success: //$($(this)).closest("tr").remove() 
-        //console.log(tr)
 
-        function(){
-            var row = tr.parentNode.parentNode;
-            row.parentNode.removeChild(row);
-        }
+    if(confirm("Are you sure you want to delete?"))
+    {
 
-    })
+        csrftoken = getCookie('csrftoken');
+        event.preventDefault();
+        /*var id = $(this).attr('id');
+        var tr = $(this).closest('tr');*/
+        $.ajax({
+            type: "DELETE",
+            url: id, 
+            headers:{
+                    "X-CSRFToken": csrftoken
+                    },
+            success: //$($(this)).closest("tr").remove() 
+            //console.log(tr)
+
+            function(){
+                var row = tr.parentNode.parentNode;
+                row.parentNode.removeChild(row);
+            }
+
+        })
+
+    }
+
+    else
+    {
+        return false;
+    }
+    
 }
